🩹 fix ✨ update  ConstantContainer, LogicHelper, TimerManager,
# ConstantContainer🩹 fix
- CONSOLE_HEIGHT 50 -> 51로 변경

# LogicHelper🩹 fix
- GetTimeSecond : 기존에 정확도가 떨어져서 현재 시간을 double로 반환함.

# TimerManager🩹 fix ✨ update
- 정확도 향상 위해 시간 관련 함수 double형으로 쓰도록 변경
- AudioPlayer의 오디오 루프 재생 기능을 위한 Tick 함숨포인터 추가

+++ textRPG/Source/Game/LogicHelper.h
@@ -36,11 +36,11 @@ public: //콘솔 관련
 public: // 시간
 
 
-	static float GetTimeSecond()
+	static double GetTimeSecond()
 	{
 		auto Now = std::chrono::system_clock::now();
 		long long Milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(Now.time_since_epoch()).count();
-		return Milliseconds / 1000.0f;
+		return Milliseconds / 1000.0;
 	}
 
 	// MiliSeconds 시간 동안 해당 쓰레드 정지
